+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+- fix [#374](https://github.com/JLLeitschuh/ktlint-gradle/issues/374): fix pre-commit hook path handling for projects in subdirectories [#996](https://github.com/JLLeitschuh/ktlint-gradle/pull/996)
+
 ## [14.0.1] - 2025-11-10
 
 - Update build to work with gradle 9.1 and Java 25 [#962](https://github.com/JLLeitschuh/ktlint-gradle/pull/962)