+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+## [14.1.0] - 2026-03-05
+
 - Add support for Android projects with new DSL i.e. built-in Kotlin [#1016](https://github.com/JLLeitschuh/ktlint-gradle/pull/1016)
 
 ## [14.0.1] - 2025-11-10
@@ -790,7 +792,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - Renamed task with name `ktlint` to `ktlintCheck` (#3)
 - Renamed tasks with names `ktlint[source set name]` to `ktlint[source set name]Check` (#3)
 
-[unreleased]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v14.0.1...HEAD
+[unreleased]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v14.1.0...HEAD
+[14.1.0]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v14.0.1...v14.1.0
 [14.0.1]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v14.0.0...v14.0.1
 [14.0.0]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v13.1.0...v14.0.0
 [13.1.0]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v13.0.0...v13.1.0